dogs vs cats/requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.20.0
numpy
pillow
